corro-client: don't depend on corro-agent
This pulls way too much excessive stuff into the proxy and causes
it to fail automatic security audit.

Instead, move the utils that have to be shared between the agent and
the client into a separate small crate.
