+++ crates/corro-agent/src/agent/run_root.rs
@@ -94,7 +94,7 @@ async fn run(agent: Agent, opts: AgentOptions, pconf: PerfConfig) -> eyre::Resul
     util::initialise_foca(&agent).await;
 
     // Load schema from paths
-    let stmts = util::read_files_from_paths(&agent.config().db.schema_paths).await?;
+    let stmts = corro_utils::read_files_from_paths(&agent.config().db.schema_paths).await?;
     if !stmts.is_empty() {
         if let Err(e) = execute_schema(&agent, stmts).await {
             error!("could not execute schema: {e}");
