+++ crates/corro-client/Cargo.toml
@@ -10,7 +10,7 @@ license = "MIT"
 [dependencies]
 bytes = { workspace = true }
 corro-api-types = { version = "0.1.0-alpha.1", path = "../corro-api-types" }
-corro-agent = { path = "../corro-agent" }
+corro-utils = { path = "../corro-utils" }
 futures = { workspace = true }
 hickory-resolver = { workspace = true }
 http = { workspace = true }
